test: move to multisig-join test file
